requirements.txt
asgiref==3.8.1
asttokens==2.4.1
black==24.3.0
click==8.1.7
decorator==5.1.1
Django==5.0.4
django-extensions==3.2.3
django-filter==24.2
django-shell-plus==1.1.7
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
executing==2.0.1
ipython==8.23.0
jedi==0.19.1
Markdown==3.6
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
packaging==24.0
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psycopg==3.1.18
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
PyJWT==2.8.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.2
typing_extensions==4.11.0
wcwidth==0.2.13
